# src/api/docs/swagger.yaml
basePath: /
definitions:
  handler.APIResponse:
    properties:
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ListAPIResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/receipt.Receipt'
        type: array
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ReceiptAPIResponse:
    properties:
      item:
        $ref: '#/definitions/receipt.Receipt'
        type: object
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  media.Media:
    properties:
      format:
        description: format (read only) (required)
        example: image/png
        type: string
      link:
        description: media link (read only) (required)
        example: ac146571eb55d9cbde1a886dbc1f85d2/event.png
        type: string
    type: object
  receipt.Receipt:
    properties:
      category:
        type: string
      cooking_time:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      media:
        $ref: '#/definitions/media.Media'
        type: object
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  services.CreateReceiptRequest:
    properties:
      category:
        maxLength: 255
        minLength: 3
        type: string
      cooking_time:
        minimum: 1
        type: integer
      description:
        maxLength: 255
        minLength: 3
        type: string
      name:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - category
    - cooking_time
    - description
    - name
    type: object
...
  services.UpdateReceiptRequest:
    properties:
      category:
        maxLength: 255
        minLength: 3
        type: string
      cooking_time:
        minimum: 1
        type: integer
      description:
        maxLength: 255
        minLength: 3
        type: string
      name:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - category
    - cooking_time
    - description
    - name
    type: object
...
host: api.food.test
info:
  contact: {}
  description: This is Swagger docs for Food API
  license: {}
  title: Food API
  version: "1.0"
paths:
...
  /v1/receipts/:
    get:
      description: find receipts by params
      parameters:
      - description: category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get receipts
      tags:
      - receipts
    post:
      parameters:
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.CreateReceiptRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create receipt
      tags:
      - receipts
  /v1/receipts/{id}:
    delete:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete receipt
      tags:
      - receipts
    put:
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.UpdateReceiptRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update receipt
      tags:
      - receipts
...
  /v1/receipts/{id}/media:
    post:
      consumes:
      - multipart/form-data
      description: upload a receipt media
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: integer
      - description: Media file
        in: formData
        name: media_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a receipt media
      tags:
      - receipts
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"